apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

# Settings / validation
pydantic-settings==2.7.1

# Database
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.14.0

# Auth / security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis cache + rate limiting store (optional)
redis==5.2.1

# Async tasks
celery==5.4.0

# OAuth + HTTP
authlib==1.3.2
httpx==0.27.2

# MFA (TOTP)
pyotp==2.9.0

# EmailStr support
email-validator==2.2.0

# Scraping + parsing + images
beautifulsoup4==4.12.3
lxml==5.3.0
pillow==10.4.0

# S3-compatible storage
boto3==1.35.90


